requirements.txt
PyYAML~=5.4.1
scapy~=2.4.4